fixed a bug trying to set an integer timeout on apache as a long

+++ src/main/java/bad/robot/http/configuration/HttpTimeout.java
@@ -36,6 +36,6 @@ public class HttpTimeout extends AbstractValueType<Duration> implements Configur
 
     @Override
     public void applyTo(Configurable configurable) {
-        configurable.setTo(value.inMillis());
+        configurable.setTo((int) value.inMillis());
     }
 }
